+++ src/main/webapp/app/exam/overview/summary/exercises/file-upload-exam-summary/file-upload-exam-summary.component.ts
@@ -1,7 +1,7 @@
 import { Component, Input } from '@angular/core';
 import { FileUploadSubmission } from 'app/entities/file-upload-submission.model';
 import { Exercise } from 'app/entities/exercise.model';
-import { FileUploadSubmissionComponent } from 'app/file-upload/overview/file-upload-submission.component';
+import { FileUploadSubmissionComponent } from 'app/fileupload/overview/file-upload-submission.component';
 
 @Component({
     selector: 'jhi-file-upload-exam-summary',
