+++ src/main/webapp/app/exercise/dashboards/tutor/exercise-assessment-dashboard.component.ts
@@ -22,7 +22,7 @@ import { Observable, of } from 'rxjs';
 import { finalize, map } from 'rxjs/operators';
 import { StatsForDashboard } from 'app/course/dashboards/stats-for-dashboard.model';
 import { TranslateService } from '@ngx-translate/core';
-import { FileUploadSubmissionService } from 'app/file-upload/overview/file-upload-submission.service';
+import { FileUploadSubmissionService } from 'app/fileupload/overview/file-upload-submission.service';
 import { FileUploadExercise } from 'app/entities/file-upload-exercise.model';
 import { ProgrammingExercise } from 'app/entities/programming/programming-exercise.model';
 import { ProgrammingSubmissionService } from 'app/programming/overview/programming-submission.service';
