+++ src/main/webapp/app/fileupload/manage/assess/file-upload-assessment.route.ts
@@ -6,7 +6,7 @@ import { Authority } from 'app/shared/constants/authority.constants';
 export const routes: Routes = [
     {
         path: 'assessment',
-        loadComponent: () => import('app/file-upload/manage/assess/file-upload-assessment.component').then((m) => m.FileUploadAssessmentComponent),
+        loadComponent: () => import('app/fileupload/manage/assess/file-upload-assessment.component').then((m) => m.FileUploadAssessmentComponent),
         data: {
             authorities: [Authority.ADMIN, Authority.INSTRUCTOR, Authority.EDITOR, Authority.TA],
             pageTitle: 'artemisApp.fileUploadExercise.home.title',
@@ -15,7 +15,7 @@ export const routes: Routes = [
     },
     {
         path: 'assessments/:resultId',
-        loadComponent: () => import('app/file-upload/manage/assess/file-upload-assessment.component').then((m) => m.FileUploadAssessmentComponent),
+        loadComponent: () => import('app/fileupload/manage/assess/file-upload-assessment.component').then((m) => m.FileUploadAssessmentComponent),
         data: {
             authorities: [Authority.ADMIN, Authority.INSTRUCTOR, Authority.EDITOR, Authority.TA],
             pageTitle: 'artemisApp.fileUploadExercise.home.title',
