+++ src/main/webapp/app/fileupload/manage/file-upload-exercise-detail.component.ts
