+++ src/main/webapp/app/fileupload/manage/file-upload-exercise-management-resolve.service.ts
@@ -1,5 +1,5 @@
 import { FileUploadExercise } from 'app/entities/file-upload-exercise.model';
-import { FileUploadExerciseService } from 'app/file-upload/manage/file-upload-exercise.service';
+import { FileUploadExerciseService } from 'app/fileupload/manage/file-upload-exercise.service';
 import { CourseManagementService } from 'app/course/manage/course-management.service';
 import { ExerciseGroupService } from 'app/exam/manage/exercise-groups/exercise-group.service';
 import { ExerciseGroup } from 'app/entities/exercise-group.model';
