+++ src/main/webapp/app/fileupload/manage/file-upload-exercise-update.component.ts
