+++ src/main/webapp/app/fileupload/manage/file-upload-exercise.component.html
