+++ src/test/javascript/spec/component/exam-exercise-row-buttons/exam-exercise-row-buttons.component.spec.ts
@@ -9,7 +9,7 @@ import { ModelingExerciseService } from 'app/modeling/manage/modeling-exercise.s
 import { TextExerciseService } from 'app/text/manage/text-exercise/text-exercise.service';
 import { ModelingExercise } from 'app/entities/modeling-exercise.model';
 import { QuizExerciseService } from 'app/quiz/manage/quiz-exercise.service';
-import { FileUploadExerciseService } from 'app/file-upload/manage/file-upload-exercise.service';
+import { FileUploadExerciseService } from 'app/fileupload/manage/file-upload-exercise.service';
 import { QuizExercise } from 'app/entities/quiz/quiz-exercise.model';
 import { FileUploadExercise } from 'app/entities/file-upload-exercise.model';
 import { of, throwError } from 'rxjs';
