+++ src/test/javascript/spec/component/exam/participate/summary/exercises/file-upload-exam-summary.component.spec.ts
@@ -2,7 +2,7 @@ import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { FileUploadExamSummaryComponent } from 'app/exam/overview/summary/exercises/file-upload-exam-summary/file-upload-exam-summary.component';
 import { FileUploadSubmission } from 'app/entities/file-upload-submission.model';
 import { By } from '@angular/platform-browser';
-import { FileUploadSubmissionComponent } from 'app/file-upload/overview/file-upload-submission.component';
+import { FileUploadSubmissionComponent } from 'app/fileupload/overview/file-upload-submission.component';
 import { FileUploadExercise } from 'app/entities/file-upload-exercise.model';
 import { MockActivatedRoute } from '../../../../../helpers/mocks/activated-route/mock-activated-route';
 import { ActivatedRoute } from '@angular/router';
