+++ src/test/javascript/spec/component/file-upload-assessment/file-upload-assessment.component.spec.ts
@@ -7,15 +7,15 @@ import dayjs from 'dayjs/esm';
 import { MockSyncStorage } from '../../helpers/mocks/service/mock-sync-storage.service';
 import { MockComponent, MockPipe } from 'ng-mocks';
 import { ActivatedRoute, Router, RouterModule } from '@angular/router';
-import { FileUploadAssessmentComponent } from 'app/file-upload/manage/assess/file-upload-assessment.component';
+import { FileUploadAssessmentComponent } from 'app/fileupload/manage/assess/file-upload-assessment.component';
 import { DebugElement } from '@angular/core';
 import { MockAccountService } from '../../helpers/mocks/service/mock-account.service';
 import { LocalStorageService, SessionStorageService } from 'ngx-webstorage';
 import { ComplaintService, EntityResponseType } from 'app/assessment/shared/complaint.service';
 import { MockComplaintService } from '../../helpers/mocks/service/mock-complaint.service';
 import { FileUploadExercise } from 'app/entities/file-upload-exercise.model';
-import { FileUploadSubmissionService } from 'app/file-upload/overview/file-upload-submission.service';
-import { FileUploadAssessmentService } from 'app/file-upload/manage/assess/file-upload-assessment.service';
+import { FileUploadSubmissionService } from 'app/fileupload/overview/file-upload-submission.service';
+import { FileUploadAssessmentService } from 'app/fileupload/manage/assess/file-upload-assessment.service';
 import { AssessmentAfterComplaint, ComplaintsForTutorComponent } from 'app/assessment/manage/complaints-for-tutor/complaints-for-tutor.component';
 import { UpdatingResultComponent } from 'app/exercise/result/updating-result.component';
 import { FileUploadSubmission } from 'app/entities/file-upload-submission.model';
@@ -23,7 +23,7 @@ import { getFirstResult, setLatestSubmissionResult, SubmissionExerciseType, Subm
 import { ExerciseType } from 'app/entities/exercise.model';
 import { AssessmentType } from 'app/entities/assessment-type.model';
 import { Result } from 'app/entities/result.model';
-import { routes } from 'app/file-upload/manage/assess/file-upload-assessment.route';
+import { routes } from 'app/fileupload/manage/assess/file-upload-assessment.route';
 import { Participation, ParticipationType } from 'app/entities/participation/participation.model';
 import { CollapsableAssessmentInstructionsComponent } from 'app/assessment/manage/assessment-instructions/collapsable-assessment-instructions/collapsable-assessment-instructions.component';
 import { AssessmentInstructionsComponent } from 'app/assessment/manage/assessment-instructions/assessment-instructions/assessment-instructions.component';
